newticketseller v3/pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot==20.7",
    "sqlalchemy==2.0.23",
]
